liberty/desktop/requirements.txt
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0
PyOpenGL>=3.1.0
numpy>=1.21.0
Pillow>=9.0.0
